# internal/api/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: Aether Notebooks API — collaborative SQL/data notebook platform
  title: Aether API
  version: 1.0.0
paths:
...
  /mcp-servers:
    get:
      description: List all MCP servers for the organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List MCP servers
      tags:
      - mcp-servers
    post:
      consumes:
      - application/json
      description: Create a new MCP server configuration
      parameters:
      - description: MCP server details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an MCP server
      tags:
      - mcp-servers
  /mcp-servers/{id}:
    delete:
      description: Delete an MCP server configuration
      parameters:
      - description: MCP Server ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an MCP server
      tags:
      - mcp-servers
    get:
      description: Get an MCP server by ID
      parameters:
      - description: MCP Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an MCP server
      tags:
      - mcp-servers
    put:
      consumes:
      - application/json
      description: Update an MCP server configuration
      parameters:
      - description: MCP Server ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP server updates
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an MCP server
      tags:
      - mcp-servers
  /mcp-servers/{id}/test:
    post:
      description: Test connectivity to an MCP server
      parameters:
      - description: MCP Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Test an MCP server
      tags:
      - mcp-servers
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"